trying a new arch

+++ train.py
@@ -53,10 +53,12 @@ val_loader = DataLoader(
 
 # CLIP image encoder (ViT-B/16)
 clip_img_model, preprocess = clip.load("ViT-B/16", device=args.device)
+clip_img_model = clip_img_model.float()
 clip_img_model.eval()
 
 # CLIP text encoder (ViT-B/32)
 clip_txt_model, _ = clip.load("ViT-B/32", device=args.device)
+clip_txt_model = clip_txt_model.float()
 clip_txt_model.eval()
 
 for p in clip_img_model.parameters():
@@ -76,32 +78,29 @@ optimizer = torch.optim.AdamW(decoder.parameters(), lr=args.learning_rate)
 
 def extract_clip_image_features(images):
     with torch.no_grad():
-        images = images.to(dtype=clip_img_model.dtype)
-
         x = clip_img_model.visual.conv1(images)          # [B, C, H/16, W/16]
         x = x.reshape(x.shape[0], x.shape[1], -1)        # [B, C, N]
         x = x.permute(0, 2, 1)                            # [B, N, C]
 
-        cls_token = clip_img_model.visual.class_embedding.to(x.dtype)
+        cls_token = clip_img_model.visual.class_embedding
         cls_token = cls_token.to(x.dtype)
         cls_token = cls_token.expand(x.shape[0], 1, -1)
 
-        x = torch.cat([cls_token, x], dim=1)
-        x = x + clip_img_model.visual.positional_embedding.to(x.dtype)
+        x = torch.cat([cls_token, x], dim=1)              # [B, N+1, C]
+        x = x + clip_img_model.visual.positional_embedding
         x = clip_img_model.visual.ln_pre(x)
 
-        x = x.permute(1, 0, 2)
+        x = x.permute(1, 0, 2)                             # [N+1, B, C]
         x = clip_img_model.visual.transformer(x)
-        x = x.permute(1, 0, 2)
+        x = x.permute(1, 0, 2)                             # [B, N+1, C]
 
         x = clip_img_model.visual.ln_post(x)
 
-        patch_tokens = x[:, 1:, :]
+        patch_tokens = x[:, 1:, :]                         # remove CLS
         B, N, C = patch_tokens.shape
         H = W = int(N ** 0.5)
-
         patch_tokens = patch_tokens.permute(0, 2, 1).contiguous()
-        patch_tokens = patch_tokens.view(B, C, H, W)
+        patch_tokens = patch_tokens.view(B, C, H, W)      # [B, C, H, W]
 
     return patch_tokens
